+++ src/vkgs_project_reader.h
@@ -58,6 +58,7 @@ private:
 
   // Sub-helper functions for meshes
   static void loadMeshAssets(const json&                             data,
+                             int                                     fileVersion,
                              const std::string&                      projectPath,
                              std::map<int, std::shared_ptr<MeshVk>>& assetIdToMesh,
                              GaussianSplattingUI*                    ui);
@@ -66,7 +67,7 @@ private:
                                 const std::map<int, std::shared_ptr<MeshVk>>& assetIdToMesh,
                                 GaussianSplattingUI*                          ui);
 
-  static void loadEnvironment(const json& data, const std::string& projectPath, GaussianSplattingUI* ui);
+  static void loadEnvironment(const json& data, int fileVersion, const std::string& projectPath, GaussianSplattingUI* ui);
   static void loadSettings(const json& data, GaussianSplattingUI* ui);
   static void loadTonemapping(const json& data, GaussianSplattingUI* ui);
 };